requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
binaryornot==0.4.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.0
daphne==4.1.2
Django==5.0.6
django-browser-reload==1.17.0
django-tailwind==3.8.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.1.1
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langsmith==0.2.10
lxml==5.3.0
mariadb==1.1.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.6
networkx==3.4.2
numpy==2.2.1
orjson==3.10.14
packaging==24.2
pgvector==0.3.6
pillow==11.1.0
propcache==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
pyOpenSSL==24.3.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-slugify==8.0.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
service-identity==24.2.0
setuptools==75.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.0
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.66.4
transformers==4.48.0
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.18.3
zope.interface==7.2
